Add pipeline_communicate() generator
This generator will be used as the "send" operation during execute()
step when the connection is in pipeline mode. It can consume results or
send queries depending on socket read/write ready state.

Added test script is taken from PostgreSQL sources. It demonstrates the
use of pipeline mode where query-send and results-fetch steps are
interleaved without any sync point emitted. (In this test, this works
because the output buffer gets full.) The test writes data to logs.
Typically, we'd get:

    enter pipeline
    sent begin: BEGIN TRANSACTION
    sent drop table: DROP TABLE IF EXISTS pq_pipeline_demo
    sent create table: CREATE UNLOGGED TABLE pq_pipeline_demo( id serial primary key, itemno integer, int8filler int8)
    sent prepare: INSERT INTO pq_pipeline_demo(itemno, int8filler) VALUES ($1, $2)
    sent row 9999
    sent row 9998
    sent row 9997
    sent row 9996
    sent row 9995
    sent row 9994
    sent row 9993
    sent row 9992
    sent row 9991
    ...
    sent row 9690
    sent row 9689
    sent row 9688
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    ...
    got COMMAND_OK results
    got COMMAND_OK results
    sent row 9687
    sent row 9686
    sent row 9685
    sent row 9684
    sent row 9683
    ...
    sent row 2
    sent row 1
    sent row 0
    sent COMMIT
    pipeline sync sent
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    got COMMAND_OK results
    ...
    got COMMAND_OK resultsd
    got COMMAND_OK results
    got PIPELINE_SYNC results
    exit pipeline

We can see that commands are sent, until the output buffer is full (the
connection is then Read-ready only), then results are fetched, until
more commands can be sent, and the cycle repeat.