Add support for pipeline mode in execute()/fetch*()
BaseConnection gets a _pipeline attribute which stores the currently
active pipeline in order to get access to the queue of commands to send
and the queue of results to process.

The command queue simply contains Callable[[], None], which are built
from partial application of pgconn.send_*() methods.

The queue of results to process either contains a None, when respective
command returns no tuple, or a tuple with respective cursor and query
information needed to maintain automatic prepared statement.

Everywhere we run the execute() generator in non-pipeline mode, we now
simply enqueue items in the pipeline queues. Then we run
pipeline_communicate(), through the _communicate_gen() method
of BasePipeline, in BaseCursor._execute(many)_gen().

Since pipeline_communicate() may not fetch all results, we need a
dedicated fetch (forced) step upon call to cursor.fetch*(); this is done
by Cursor._fetch_pipeline() called in fetch*() methods. This usually
calls PQsendFlushRequest() in order to avoid blocking on PQgetResult().

Similarly, at exit of pipeline mode, we force results fetch after
sending any pending commands (e.g. commands not emitted through an
execute() call).

The pipeline-demo.py test script is updated to include examples using
the high-level API. This only works with PSYCOPG_IMPL=python because we
monkeypatch the pgconn attribute of the connection.