feat: add waiting functions using anyio
We wrap anyio.wait_socket_{readable,writable}() functions to accommodate
for types constraints on anyio side that seem too rigid for our case.

In tests, we define 'wait_{conn_,}_async' fixtures that will pick either
asyncio or anyio waiting functions depending on the value of
'anyio_backend' fixture, which is redefined in test_waiting.py to get
parametrized with asyncio and trio backends. Accordingly, we now depend
on anyio[trio] in tests.