Move threads start/stop of ConnectionPool at context manager level
We move threads creation and start from __init__() to __enter__().
Threads stop is now only handled in close(), called by __exit__(), and
the __del__ method is no longer useful.

The _closed attribute is handled in __enter__()/__exit__() now; thus it
is set to True in __init__() (in base class, thus also updating the
AsyncConnectionPool).