Use anyio in async pool [skip ci] (WIP)
The task receiver stream is cloned for each worker so that each clone
can be closed by their respective worker.

To terminate workers, we now close the sending stream; this will emit an
EndOfStream, and thus a StopIteration, on the receiving end. Drop the
StopWorker task class no longer useful now.

TODO:
* understand how workers work (why are there multiple of them? is it
  safe that they share the same receive stream?)
* should run_task() be async, and call _task_sender.send() instead of
  send_nowait()?