feat: add support for AnyIO in copy()
We add AnyIOLibpqWriter class that uses an AnyIO task group to hold the
worker task and memory object streams for communication.

We add a _copywritercls attribute on AsyncConnection base class and
AnyIOConnection so that the AsyncCopy instance can use the correct
writer depending on the connection async backend.

The finish() method of AsyncCopy and AsyncWriter now take full exception
information so that we can use it when exiting the AnyIO task group.

We ignore an invalid error (raised by a deprecation warning during test)
about usage of the 'loop' parameter in asyncio API that is due to Python
bug as mentioned in comment.