feat: possibly enter pipeline mode upon transaction
The pipeline wraps the transaction, meaning that a Sync will be sent
before transaction's commit or rollback. Yet we emit a Sync at the
beginning of transaction block to make sure the connection status is
right (this breaks tests otherwise and seems easier to handle, at least
as a first step).

Handling for pipeline mode availability is done through new NullPipeline
classes.