feat: add Connection.pause_pipeline() context manager
This would be useful when the connection is in pipeline mode (possibly
set from an outer/out-of-control scope) and one wants to perform
operations that are not supported in pipeline mode (e.g. COPY).

The context manager is a no-op if the connection is not in pipeline to
compose well with outer scopes.