feat: add non-blocking cancellation
We possibly (libpq >= 16) use the non-blocking libpq cancellation API in
connection.cancel(). In AsyncConnection, an acancel() async method is
introduced to preserve compatibility whereas the cancel() method used
the old libpq API, which is suitable for use in a signal handler (as in
CTRL+C handling).

The logic of Copy termination, in finish(), is reworked so that
connection cancellation is invoked from there directly instead of from
_end_copy_out_gen() as we cannot call async code from the generator.

CTRL+C handling also use the non-blocking cancellation.